fix: csp directives
