+++ assets/sso_complete.html.tmpl
@@ -9,7 +9,7 @@
         crossorigin="anonymous"
     ></script>
     <meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0">
-    <style>
+    <style nonce="{{.Nonce}}">
         body {
             font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
             display: flex;
