+++ assets/sso_wait.html.tmpl
@@ -4,7 +4,7 @@
 	<meta charset="UTF-8">
 	<title>Mattermost</title>
 	<meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0">
-	<style>
+	<style nonce="{{.Nonce}}">
 		body {
 			font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
 			display: flex;
